requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
openpyxl==3.1.2
python-docx==1.1.0
PyPDF2==3.0.1
Pillow==10.2.0
Jinja2==3.1.3
aiofiles==23.2.1
PyYAML==6.0.1
docx2pdf==0.1.8  # For PDF conversion using Microsoft Word (Windows)

